+++ server/src/ui/drawer.js
@@ -52,14 +52,14 @@ function drawer(req, perms) {
           }
         : BLOCKED,
     pokestops:
-      (perms.pokestops || perms.lures || perms.quests || perms.invasions) &&
+      (perms.pokestops || perms.quests || perms.invasions || perms.lures) &&
       state.db.models.Pokestop
         ? {
             allPokestops: perms.pokestops || BLOCKED,
-            lures: perms.lures || BLOCKED,
-            eventStops: perms.eventStops || BLOCKED,
             quests: perms.quests || BLOCKED,
             invasions: perms.invasions || BLOCKED,
+            eventStops: perms.eventStops || BLOCKED,
+            lures: perms.lures || BLOCKED,
             arEligible: perms.pokestops || BLOCKED,
           }
         : BLOCKED,