+++ src/features/pokestop/PokestopTile.jsx
@@ -12,6 +12,10 @@ import { useManualPopupTracker } from '@hooks/useManualPopupTracker'
 import { TooltipWrapper } from '@components/ToolTipWrapper'
 
 import { PokestopPopup } from './PokestopPopup'
+import {
+  INCIDENT_DISPLAY_TYPES,
+  getPokestopIncidentState,
+} from './incidentPriority'
 import { usePokestopMarker } from './usePokestopMarker'
 
 /**
@@ -22,6 +26,16 @@ import { usePokestopMarker } from './usePokestopMarker'
 const BasePokestopTile = (pokestop) => {
   const [stateChange, setStateChange] = React.useState(false)
   const [markerRef, setMarkerRef] = React.useState(null)
+  const ts = Date.now() / 1000
+  const incidentState = getPokestopIncidentState({
+    events: pokestop.events,
+    invasions: pokestop.invasions,
+    showcase_expiry: pokestop.showcase_expiry,
+    incident_blocker_display_type: pokestop.incident_blocker_display_type,
+    incident_blocker_expire_timestamp:
+      pokestop.incident_blocker_expire_timestamp,
+    ts,
+  })
   const hasRoutes = useRouteStore(
     React.useCallback(
       (state) =>
@@ -37,111 +51,143 @@ const BasePokestopTile = (pokestop) => {
   const selectPoi = useRouteStore((s) => s.selectPoi)
 
   const [
-    hasLure,
-    hasInvasion,
-    hasQuest,
-    hasEvent,
-    hasAllStops,
-    showTimer,
+    canShowLures,
+    canShowInvasions,
+    canShowQuests,
+    canShowEvents,
+    canShowPokestops,
+    hasTimerOverride,
     interactionRangeZoom,
-    hasShowcase,
-  ] = useMemory((s) => {
-    const newTs = Date.now() / 1000
-    const { filters } = useStorage.getState()
-    const {
-      config,
-      timerList,
-      auth: { perms },
-    } = s
-    return [
-      pokestop.lure_expire_timestamp > newTs && perms.lures,
-      !!(
-        perms.invasions &&
-        pokestop.invasions?.some(
-          (invasion) =>
-            invasion.grunt_type && invasion.incident_expire_timestamp > newTs,
-        )
-      ),
-      !!(perms.quests && pokestop.quests?.length),
-      !!(
-        perms.eventStops &&
-        filters.pokestops.eventStops &&
-        pokestop.events?.some((event) => event.event_expire_timestamp > newTs)
-      ),
-      (filters.pokestops.allPokestops || pokestop.ar_scan_eligible) &&
-        perms.pokestops,
-      timerList.includes(pokestop.id),
-      config.general.interactionRangeZoom,
-      !!(perms.pokestops && pokestop.hasShowcase),
-    ]
-  }, basicEqualFn)
+  ] = useMemory(
+    (s) => [
+      !!s.auth.perms.lures,
+      !!s.auth.perms.invasions,
+      !!s.auth.perms.quests,
+      !!s.auth.perms.eventStops,
+      !!s.auth.perms.pokestops,
+      s.timerList.includes(pokestop.id),
+      s.config.general.interactionRangeZoom,
+    ],
+    basicEqualFn,
+  )
 
   const [
-    invasionTimers,
-    lureTimers,
-    eventStopTimers,
-    lureRange,
-    showcaseRange,
-    interactionRange,
+    showEventStops,
+    showAllStops,
+    showInvasionTimers,
+    showLureTimers,
+    showEventStopTimers,
+    showLureRange,
+    showShowcaseRange,
+    showInteractionRange,
     customRange,
+    zoom,
   ] = useStorage((s) => {
-    const { userSettings, zoom } = s
+    const { userSettings } = s
     return [
-      userSettings.pokestops.invasionTimers || showTimer,
-      userSettings.pokestops.lureTimers || showTimer,
-      userSettings.pokestops.eventStopTimers || showTimer,
-      !!userSettings.pokestops.lureRange && zoom >= interactionRangeZoom,
-      !!userSettings.pokestops.showcaseRange &&
-        zoom >= interactionRangeZoom &&
-        hasShowcase,
-      !!userSettings.pokestops.interactionRanges &&
-        zoom >= interactionRangeZoom,
-      zoom >= interactionRangeZoom
-        ? +userSettings.pokestops.customRange || 0
-        : 0,
+      !!s.filters.pokestops.eventStops,
+      !!s.filters.pokestops.allPokestops,
+      !!(userSettings.pokestops.invasionTimers || hasTimerOverride),
+      !!(userSettings.pokestops.lureTimers || hasTimerOverride),
+      !!(userSettings.pokestops.eventStopTimers || hasTimerOverride),
+      !!userSettings.pokestops.lureRange,
+      !!userSettings.pokestops.showcaseRange,
+      !!userSettings.pokestops.interactionRanges,
+      +userSettings.pokestops.customRange || 0,
+      s.zoom,
     ]
   }, basicEqualFn)
 
-  const timers = React.useMemo(() => {
-    const internalTimers = /** @type {number[]} */ ([])
-    if (invasionTimers && hasInvasion) {
-      pokestop.invasions.forEach((invasion) =>
-        internalTimers.push(invasion.incident_expire_timestamp),
+  const hasLure = pokestop.lure_expire_timestamp > ts && canShowLures
+  const hasQuest = !!(canShowQuests && pokestop.quests?.length)
+  const hasInvasion = !!(
+    canShowInvasions && incidentState.popupInvasions.length
+  )
+  const hasEvent = !!(
+    canShowEvents &&
+    showEventStops &&
+    incidentState.popupEvents.length
+  )
+  const visibleMarkerInvasions = canShowInvasions
+    ? incidentState.markerInvasions
+    : []
+  const visibleMarkerEvents =
+    canShowEvents && showEventStops ? incidentState.markerEvents : []
+  const hasVisibleInvasion = !!(
+    canShowInvasions && visibleMarkerInvasions.length
+  )
+  const hasVisibleEvent = !!visibleMarkerEvents.length
+  const hasVisibleShowcase = visibleMarkerEvents.some(
+    (event) =>
+      Number(event.display_type ?? 0) === INCIDENT_DISPLAY_TYPES.SHOWCASE,
+  )
+  const hasAllStops = !!(
+    (showAllStops || pokestop.ar_scan_eligible) &&
+    canShowPokestops
+  )
+  const withinRangeZoom = zoom >= interactionRangeZoom
+  const lureRange = showLureRange && withinRangeZoom
+  const showcaseRange =
+    showShowcaseRange && withinRangeZoom && hasVisibleShowcase
+  const interactionRange = showInteractionRange && withinRangeZoom
+  const renderedCustomRange = withinRangeZoom ? customRange : 0
+
+  const [refreshTimers, tooltipTimers] = React.useMemo(() => {
+    const internalRefreshTimers = [...incidentState.expiryTimestamps]
+    const internalTooltipTimers = /** @type {number[]} */ ([])
+
+    if (showInvasionTimers && hasVisibleInvasion) {
+      visibleMarkerInvasions.forEach((invasion) =>
+        internalTooltipTimers.push(invasion.incident_expire_timestamp),
       )
     }
-    if (lureTimers && hasLure) {
-      internalTimers.push(pokestop.lure_expire_timestamp)
+    if (showLureTimers && hasLure) {
+      internalRefreshTimers.push(pokestop.lure_expire_timestamp)
+      internalTooltipTimers.push(pokestop.lure_expire_timestamp)
     }
-    if (eventStopTimers && hasEvent) {
-      pokestop.events.forEach((event) => {
-        internalTimers.push(event.event_expire_timestamp)
+    if (showEventStopTimers && hasVisibleEvent) {
+      visibleMarkerEvents.forEach((event) => {
+        internalTooltipTimers.push(event.event_expire_timestamp)
       })
     }
-    return internalTimers
+
+    return [internalRefreshTimers, internalTooltipTimers]
   }, [
-    invasionTimers,
-    hasInvasion,
-    lureTimers,
+    incidentState.expiryTimestamps,
+    visibleMarkerEvents,
+    visibleMarkerInvasions,
+    showInvasionTimers,
+    hasVisibleInvasion,
+    showLureTimers,
     hasLure,
-    eventStopTimers,
-    hasEvent,
+    showEventStopTimers,
+    hasVisibleEvent,
+    pokestop.lure_expire_timestamp,
   ])
 
   useForcePopup(pokestop.id, markerRef)
-  useMarkerTimer(timers.length ? Math.min(...timers) : null, markerRef, () =>
-    setStateChange(!stateChange),
+  useMarkerTimer(
+    refreshTimers.length ? Math.min(...refreshTimers) : null,
+    markerRef,
+    () => setStateChange(!stateChange),
   )
   const handlePopupOpen = useManualPopupTracker('pokestops', pokestop.id)
 
   const icon = usePokestopMarker({
     hasQuest,
     hasLure,
-    hasInvasion,
-    hasEvent,
+    markerEvents: visibleMarkerEvents,
+    markerInvasions: visibleMarkerInvasions,
+    baseIncidentDisplay:
+      canShowEvents && showEventStops ? incidentState.baseDisplay : '',
     ...pokestop,
   })
 
-  return hasQuest || hasLure || hasInvasion || hasEvent || hasAllStops ? (
+  return hasQuest ||
+    hasLure ||
+    hasVisibleInvasion ||
+    hasVisibleEvent ||
+    hasAllStops ? (
     <Marker
       ref={setMarkerRef}
       position={[pokestop.lat, pokestop.lon]}
@@ -161,11 +207,14 @@ const BasePokestopTile = (pokestop) => {
           hasInvasion={hasInvasion}
           hasQuest={hasQuest}
           hasEvent={hasEvent}
+          popupInvasions={incidentState.popupInvasions}
+          popupEvents={incidentState.popupEvents}
+          incidentBlocker={incidentState.blocker}
           {...pokestop}
         />
       </Popup>
-      {Boolean(timers.length) && (
-        <TooltipWrapper timers={timers} offset={[0, 4]} />
+      {Boolean(tooltipTimers.length) && (
+        <TooltipWrapper timers={tooltipTimers} offset={[0, 4]} />
       )}
       {interactionRange && (
         <Circle
@@ -188,10 +237,10 @@ const BasePokestopTile = (pokestop) => {
           pathOptions={{ color: '#39a18f', weight: 1 }}
         />
       )}
-      {!!customRange && (
+      {!!renderedCustomRange && (
         <Circle
           center={[pokestop.lat, pokestop.lon]}
-          radius={customRange}
+          radius={renderedCustomRange}
           pathOptions={{ color: 'purple', weight: 0.5 }}
         />
       )}
@@ -205,18 +254,31 @@ export const PokestopTile = React.memo(
     prev.id === next.id &&
     prev.lure_expire_timestamp === next.lure_expire_timestamp &&
     prev.updated === next.updated &&
-    prev.hasShowcase === next.hasShowcase &&
+    prev.showcase_expiry === next.showcase_expiry &&
+    prev.incident_blocker_display_type === next.incident_blocker_display_type &&
+    prev.incident_blocker_expire_timestamp ===
+      next.incident_blocker_expire_timestamp &&
     prev.quests?.length === next.quests?.length &&
     (prev.quests && next.quests
       ? prev.quests.every((q, i) => q.with_ar === next.quests[i]?.with_ar)
       : true) &&
     prev.invasions?.length === next.invasions?.length &&
     (prev.invasions && next.invasions
-      ? prev.invasions?.every(
+      ? prev.invasions.every(
           (inv, i) =>
-            inv.confirmed === next?.invasions?.[i]?.confirmed &&
-            inv.grunt_type === next?.invasions?.[i]?.grunt_type,
+            inv.confirmed === next.invasions?.[i]?.confirmed &&
+            inv.grunt_type === next.invasions?.[i]?.grunt_type &&
+            inv.incident_expire_timestamp ===
+              next.invasions?.[i]?.incident_expire_timestamp,
         )
       : true) &&
-    prev.events?.length === next.events?.length,
+    prev.events?.length === next.events?.length &&
+    (prev.events && next.events
+      ? prev.events.every(
+          (event, i) =>
+            event.display_type === next.events?.[i]?.display_type &&
+            event.event_expire_timestamp ===
+              next.events?.[i]?.event_expire_timestamp,
+        )
+      : true),
 )